fix: repair 14 feature-gated compile breaks from the 2cb97410 heal — manual_let_else binding-name mangling (bound the Some-arm pattern ident instead of the outer let name: preferred_i/last/slot/rhythm/op_def/idx/next_tok/owned/prop/group/winner_cell/channel/action/online_result), 1 use_self Self<A> over-reach (event_log diff), 1 vec-with-capacity i64-len E0308 (wasm expand_shl, cast added), 1 sort_by_key non-Copy E0507 (datrie build, reverted to sort_by); all sites compiled only under non-default feature sets (the verify-gate blind spot), found by workspace --all-features sweep, now clean

+++ crates/katgpt-forward/src/set_diffusion.rs
@@ -255,7 +255,7 @@ pub fn set_diffusion_decode<F: SetCausalForwardFn>(
     let max_gen_step = gen_steps.iter().copied().max().unwrap_or(0);
 
     let mut forward_passes = 0usize;
-    let mut confidence_history = Vec::with_capacity(max_gen_step + 1);
+    let mut confidence_history = Vec::with_capacity(max_gen_step as usize + 1);
     let mut all_committed = false;
 
     // Hoisted out of both loops: gen_steps doesn't change between iterations,
